requirements.txt
Faker==6.6.2
kafka-python